9205 - 맥주 마시면서 걸어가기

+++ BaekJoon/09000/9205.cpp
@@ -11,8 +11,8 @@ int main() {
   ios::sync_with_stdio(false);
   cin.tie(nullptr);
 
-  vector<pair<int, int>> yx;
-  vector<int> memo;
+  vector<pair<int16_t, int16_t>> yx;
+  vector<int8_t> memo;
 
   auto Find = [&](int id) {
     while (id != memo[id]) {
